--- app/about/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function About() {
-  return (
-    <div>This is about page</div>
-  )
-}
-
-export default About
