--- app/about/portfolio/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function page() {
-  return (
-    <div>Portfolio</div>
-  )
-}
-
-export default page
