--- app/about/users/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Users() {
-  return (
-    <div> This is About Page Users</div>
-  )
-}
-
-export default Users
