--- app/next_js_lesson/linkes.txt
@@ -1,5 +1,0 @@
-Link
-is a component provided by Next.js that allows you to create links to navigate between pages in your application.
-
-# Its primarily used for declarative, client-side navigation, When you click on a link created with link, It prevents the default prowser full page
-refresh and fetches the new page content on the client side, resulting in a faster and smoother user experience.
