--- app/next_js_lesson/routes.txt
@@ -1,23 +1,0 @@
-Routings
-***********
-This is the directions to a particular website or page.
-
-Create Folder/
-   eg: about/page.js
-
-localhost:3000/about 
-which take you the path ie. about can be any name being prefferred.
-
-Nested routings:
-With the routing you can create a nested route by basically creating 
-a main folder to serve as a routing path NB: dont forget to add the page.tsxand add the next routes you have to get to
-eg: about/user
-
-End point
-localhost:3000/about/user
-
-Folder structure:
-about/
-   page.tsx
-   users/
-       page.tsx
